consolidated simulation functions
