+++ tests/simulations/test_zones.py
@@ -1,8 +1,6 @@
-import unittest
 import copy
 from pathlib import Path
 
-from tests import BaseTest
 from tests.simulations import BaseSimulationTest
 from src.hvac_template import HVACTemplate
 from src.epjson_handler import EPJSON
@@ -43,14 +41,14 @@ mock_zone_template = {
 test_dir = Path(__file__).parent.parent
 
 
-class TestSimulationSimple(BaseTest, BaseSimulationTest, unittest.TestCase):
+class TestSimulationSimpleZone(BaseSimulationTest):
     def setUp(self):
         return
 
     def teardown(self):
         return
 
-    @BaseTest._test_logger(doc_text="HVACTemplate:Zone:VAV w/ HW Reheat Simulation test")
+    @BaseSimulationTest._test_logger(doc_text="HVACTemplate:Zone:VAV w/ HW Reheat Simulation test")
     def test_simulation(self):
         base_file_path = str(test_dir / '..' / 'simulation' / 'ExampleFiles' /
                              'HVACTemplate-5ZoneVAVWaterCooledExpanded.epJSON')
@@ -88,20 +86,7 @@ class TestSimulationSimple(BaseTest, BaseSimulationTest, unittest.TestCase):
         test_input_file_path = self.write_file_for_testing(
             epjson=output_epjson,
             file_name='test_input_epjson.epJSON')
-        # check outputs
-        status_checks = self.perform_comparison([base_input_file_path, test_input_file_path])
-        for energy_val in status_checks['total_energy_outputs']:
-            self.assertAlmostEqual(energy_val / max(status_checks['total_energy_outputs']), 1, 2)
-        for warning in status_checks['warning_outputs']:
-            self.assertEqual(warning, max(status_checks['warning_outputs']))
-        for error in status_checks['error_outputs']:
-            self.assertEqual(error, max(status_checks['error_outputs']))
-            self.assertGreaterEqual(error, 0)
-        for status in status_checks['finished_statuses']:
-            self.assertEqual(1, status)
-        # compare epJSONs
-        comparison_results = self.compare_epjsons(base_formatted_epjson, output_epjson)
-        if comparison_results:
-            # trigger failure
-            self.assertEqual('', comparison_results, comparison_results)
+        # check outputs and compare epJSONs
+        self.perform_comparison([base_input_file_path, test_input_file_path])
+        self.compare_epjsons(base_formatted_epjson, output_epjson)
         return
