+++ tests/simulations/test_zones_output_files.py
@@ -1,9 +1,7 @@
-import unittest
 import copy
 from pathlib import Path
 from argparse import Namespace
 
-from tests import BaseTest
 from tests.simulations import BaseSimulationTest
 from src.epjson_handler import EPJSON
 from src.main import main
@@ -44,7 +42,7 @@ mock_zone_template = {
 test_dir = Path(__file__).parent.parent
 
 
-class TestSimulationFiles(BaseTest, BaseSimulationTest, unittest.TestCase):
+class TestSimulationFiles(BaseSimulationTest):
     """
     Simulation testing which writes epJSON objects to non-temporary files.
     """
@@ -55,7 +53,7 @@ class TestSimulationFiles(BaseTest, BaseSimulationTest, unittest.TestCase):
     def teardown(self):
         return
 
-    @BaseTest._test_logger(doc_text="HVACTemplate:Zone:VAV w/ HW Reheat Simulation test")
+    @BaseSimulationTest._test_logger(doc_text="HVACTemplate:Zone:VAV w/ HW Reheat Simulation test")
     def test_simulation(self):
         # initialize list of files to run
         file_run_list = []
@@ -104,20 +102,7 @@ class TestSimulationFiles(BaseTest, BaseSimulationTest, unittest.TestCase):
                 no_schema=False,
                 output_directory=output_directory))
         file_run_list.append(output_directory.joinpath(output['output_files']['expanded']))
-        # check outputs
-        status_checks = self.perform_comparison(file_run_list)
-        for energy_val in status_checks['total_energy_outputs']:
-            self.assertAlmostEqual(energy_val / max(status_checks['total_energy_outputs']), 1, 2)
-        for warning in status_checks['warning_outputs']:
-            self.assertEqual(warning, max(status_checks['warning_outputs']))
-        for error in status_checks['error_outputs']:
-            self.assertEqual(error, max(status_checks['error_outputs']))
-            self.assertGreaterEqual(error, 0)
-        for status in status_checks['finished_statuses']:
-            self.assertEqual(1, status)
-        # compare epJSONs
-        comparison_results = self.compare_epjsons(original_formatted_epjson, output['epJSON'])
-        if comparison_results:
-            # trigger failure
-            self.assertEqual('', comparison_results, comparison_results)
+        # check outputs and compare epJSONs
+        self.perform_comparison(file_run_list)
+        self.compare_epjsons(original_formatted_epjson, output['epJSON'])
         return
